cleanup reorganize
